Adding documentation for changes
