+++ src/IIS/Cake.IIS.csproj
@@ -55,6 +55,7 @@
     <Compile Include="Manager\Base\BaseManager.cs" />
     <Compile Include="Manager\Types\ApplicationManager.cs" />
     <Compile Include="Manager\Types\ApplicationPoolManager.cs" />
+    <Compile Include="Manager\Types\VirtualApplicationSettings.cs" />
     <Compile Include="Manager\Types\WebFarmManager.cs" />
     <Compile Include="Settings\ApplicationSettings.cs" />
     <Compile Include="Settings\BindingSettings.cs" />
